Mariela

+++ js/cart.js
@@ -18,7 +18,7 @@ console.log(data);
 LS = JSON.parse(localStorage.getItem("idsCarrito"));
 
 if (LS) {
-  productosId = LS;
+  productosId = LS;//array de ids
 }
 
 console.log(LS);
@@ -27,7 +27,7 @@ function arrayActual(arrayProdComprados, productosId) {
   for (const iterator of productosId) {
     aux = data.find((obj) => obj.id == iterator.id);
     if (aux) {
-      arrayProdComprados.push(aux);
+      arrayProdComprados.push(aux);//array de objetos ya parseados
     }
   }
   return arrayProdComprados;
@@ -37,6 +37,7 @@ console.log(arrayProdComprados);
 //renderCardsCart(arrayActual(arrayProdComprados, productosId), contenedorCards);
 
 //CLICKS EN EL CONTENEDOR
+
 contenedorCards.addEventListener("click", (evento) => {
   let stockact = document.querySelector(`p[data-stockact="${idCaptado}"]`);
   let subtotal = document.querySelector(`h4[data-subtotal="${idCaptado}"]`);
@@ -44,7 +45,7 @@ contenedorCards.addEventListener("click", (evento) => {
   if (evento.target.dataset.menos) {
     idCaptado = evento.target.dataset.menos;
     inputCaptado = document.querySelector(`input[data-idmod="${idCaptado}"]`);
-    let auxi = data.find((obj) => obj.id == idCaptado);
+    let auxi = data.find((obj) => obj.id == idCaptado); 
 
     //posicion en el Array de LS
     let pos = productosId.findIndex((obj) => obj.id == idCaptado);
@@ -61,6 +62,7 @@ contenedorCards.addEventListener("click", (evento) => {
       localStorage.setItem("idsCarrito", JSON.stringify(productosId));
       inputCaptado.setAttribute("value", `${aux}`);
       subtotal.innerHTML = `$${aux * auxi.precio_venta}`;
+      stockact.innerHTML = `Hay actualmente ${auxi.stock} unidades disponibles`
       console.log(inputCaptado);
     }
   }
@@ -77,25 +79,25 @@ contenedorCards.addEventListener("click", (evento) => {
     //posicion en el Array de LS
     let pos = productosId.findIndex((obj) => obj.id == idCaptado);
     let auxi = data.find((obj) => obj.id == idCaptado);
-    let stock = data.find((obj) => obj.id == idCaptado).stock;
-    aux = productosId.find((obj) => obj.id == idCaptado).cantidad + 1;
+    aux = productosId.find((obj) => obj.id == idCaptado).cantidad + 1;//CANTIDAD DEL LS
 
-    if (auxi.stock > 0 && auxi.stock <= 5) {
+    //ALERTA DE ULTIMAS UNIDADES
+   /*  if (auxi.stock > 0 && auxi.stock <= 5) {
       stockact.className = " text-red-800";
-      stockact.innerHTML = "Solo quedan las últimas!! ";
-    }
-
+      stockact.innerHTML = `Solo quedan las últimas ${auxi.stock} unidades!! `;
+    }else{stockact.innerHTML = `Hay actualmente ${obj.stock} unidades disponibles`}
+ */
     //Logica de Subtotal Tarjeta
     if (aux >= 0 && aux <= auxi.stock) {
       productosId[pos].cantidad++;
       localStorage.setItem("idsCarrito", JSON.stringify(productosId));
       inputCaptado.setAttribute("value", `${aux}`);
-
       subtotal.innerHTML = `$${aux * auxi.precio_venta}`;
+      stockact.innerHTML = `Hay actualmente ${auxi.stock} unidades disponibles`
       //Aviso de stock
-    } else if (inputCaptado > auxi.stock) {
+    } else if (aux > auxi.stock) {
       subtotal.innerHTML = `$${aux * auxi.precio_venta}`;
-      stockact.innerHTML = "No tenemos las unidades requeridas.";
+      stockact.innerHTML = "No tenemos en stock las unidades requeridas.";
     } else {
       subtotal.innerHTML = "$-";
     }
@@ -109,15 +111,32 @@ contenedorCards.addEventListener("click", (evento) => {
 
 });
 
-botonSend.addEventListener('click', event=>
-alert('Gracias por tu compra!! \n En breve nos contactaremos para hacerte llegar el paquete'))
+botonSend.addEventListener('click', event=>{
+
+    productosId = productosId.filter((producto) => producto.cantidad == 0);
+    console.log(productosId);
+    localStorage.setItem("idsCarrito", JSON.stringify(productosId));
+
+    alert('Gracias por tu compra!! \n En breve nos contactaremos para hacerte llegar el paquete')
+
+})
+
 
 
 botonClear.addEventListener('click', event=>{
-localStorage.clear('idsCarrito')
-alert('Gracias por tu compra!! \n En breve nos contactaremos para hacerte llegar el paquete')
+    let confirmar = confirm("Se vaciará el carrito con todos tus productos, deseas continuar?")
+   
+    if(confirmar){
+        localStorage.clear('idsCarrito')
+    }
+    
 
 
 })
 
-renderCardsCart(arrayActual(arrayProdComprados, productosId), contenedorCards)
+console.log(productosId.length);
+if(productosId.length>0){
+    renderCardsCart(arrayActual(arrayProdComprados, productosId), contenedorCards)
+}else{ contenedorCards.innerHTML="NO HAY PRODUCTOS EN TU CARRITO"}
+
+//renderCardsCart(arrayActual(arrayProdComprados, productosId),contenedorCards)
